Moved new drop down classes to separate files, added StaticDropDownMenu and FilteredDropDownMenu
