+++ Applications/Spire/Include/Spire/Ui/ColorSelectorButton.hpp
@@ -38,19 +38,19 @@ namespace Spire {
         const ColorSignal::slot_type& slot) const;
 
     protected:
-      bool eventFilter(QObject* watched, QEvent* event) override;
-      void focusOutEvent(QFocusEvent* event) override;
-      void keyPressEvent(QKeyEvent* event) override;
-      void mousePressEvent(QMouseEvent* event) override;
+      //bool eventFilter(QObject* watched, QEvent* event) override;
+      //void focusOutEvent(QFocusEvent* event) override;
+      //void keyPressEvent(QKeyEvent* event) override;
+      //void mousePressEvent(QMouseEvent* event) override;
       void paintEvent(QPaintEvent* event) override;
 
     private:
       mutable ColorSignal m_color_signal;
       QColor m_current_color;
-      ColorSelectorDropDown* m_dropdown;
+      ColorSelectorDropDown* m_selector_widget;
 
-      void hide_dropdown();
-      void move_color_dropdown();
+      //void hide_dropdown();
+      //void move_color_dropdown();
       void on_color_selected(const QColor& color);
   };
 }
