+++ Applications/Spire/Include/Spire/Ui/DropDownWindow.hpp
@@ -1,19 +1,30 @@
 #ifndef SPIRE_DROP_DOWN_WINDOW_HPP
 #define SPIRE_DROP_DOWN_WINDOW_HPP
 #include <QWidget>
+#include "Spire/Ui/DropShadow.hpp"
 
 namespace Spire {
 
-  //! Represents a drop down window for use with a button.
   class DropDownWindow : public QWidget {
     public:
 
-      //! Constructs a DropDownWindow.
-      /*!
-        \param parent The parent widget that the DropDownWindow will be shown
-                      relative to.
-      */
-      explicit DropDownWindow(QWidget* parent = nullptr);
+      explicit DropDownWindow(bool is_click_activated,
+        QWidget* parent = nullptr);
+
+      void set_widget(QWidget* widget);
+
+    protected:
+      bool event(QEvent* event) override;
+      bool eventFilter(QObject* watched, QEvent* event) override;
+      void keyPressEvent(QKeyEvent* event) override;
+      void paintEvent(QPaintEvent* event) override;
+
+    private:
+      QWidget* m_widget;
+      DropShadow* m_shadow;
+      bool m_is_click_activated;
+
+      void move_to_parent();
   };
 }
