+++ Applications/Spire/Source/Scratch/main.cpp
@@ -3,430 +3,11 @@
 
 using namespace Spire;
 
-#include <QHBoxLayout>
-#include <QImage>
-#include <QKeyEvent>
 #include <QLabel>
-#include <QPainter>
-#include <QScrollBar>
 #include <QWidget>
 #include "Spire/Spire/Dimensions.hpp"
-#include "Spire/Ui/CustomQtVariants.hpp"
-#include "Spire/Ui/DropShadow.hpp"
-#include "Spire/Ui/ScrollArea.hpp"
-#include "Spire/Ui/Ui.hpp"
-
-class DropDownWindow : public QWidget {
-  public:
-
-    DropDownWindow(QWidget* parent = nullptr)
-        : QWidget(parent, Qt::Tool | Qt::FramelessWindowHint),
-          m_widget(nullptr) {
-      setAttribute(Qt::WA_ShowWithoutActivating);
-      setAttribute(Qt::WA_TranslucentBackground);
-      m_shadow = new DropShadow(true, false, this);
-      auto layout = new QHBoxLayout(this);
-      layout->setContentsMargins(scale_width(1), scale_height(1),
-        scale_width(1), scale_height(1));
-      parent->installEventFilter(this);
-      parent->window()->installEventFilter(this);
-    }
-
-    void set_widget(QWidget* widget) {
-      if(m_widget != nullptr) {
-        return;
-      }
-      m_widget = widget;
-      m_widget->installEventFilter(this);
-      m_widget->setFocusProxy(this);
-      layout()->addWidget(m_widget);
-    }
-
-  protected:
-    bool event(QEvent* event) override {
-      if(event->type() == QEvent::WindowDeactivate &&
-          focusWidget() != nullptr &&
-          !isAncestorOf(focusWidget())) {
-        hide();
-      }
-      return QWidget::event(event);
-    }
-    bool eventFilter(QObject* watched, QEvent* event) override {
-      if(watched == m_widget) {
-        if(event->type() == QEvent::Resize) {
-          resize(m_widget->width() + 2, m_widget->height() + 2);
-        }
-      } else if(watched == parent()) {
-        if(event->type() == QEvent::Move) {
-          move_to_parent();
-        } else if(event->type() == QEvent::FocusOut &&
-            focusWidget() != nullptr &&
-            !isAncestorOf(focusWidget())) {
-          auto a = focusWidget();
-          hide();
-        } else if(event->type() == QEvent::Wheel) {
-          hide();
-        } else if(event->type() == QEvent::KeyPress) {
-          auto e = static_cast<QKeyEvent*>(event);
-          if(e->key() == Qt::Key_Escape) {
-            hide();
-          } else if(e->key() == Qt::Key_Enter || e->key() == Qt::Key_Return) {
-            if(isVisible()) {
-              hide();
-            } else {
-              move_to_parent();
-              show();
-            }
-            return true;
-          }
-        }
-      } else if(watched == static_cast<QWidget*>(parent())->window()) {
-        if(event->type() == QEvent::WindowDeactivate && !isActiveWindow()) {
-          hide();
-        } else if(event->type() == QEvent::Move) {
-          move_to_parent();
-        }
-      }
-      return QWidget::eventFilter(watched, event);
-    }
-    void keyPressEvent(QKeyEvent* event) override {
-      if(event->key() == Qt::Key_Escape) {
-        hide();
-      }
-    }
-    void paintEvent(QPaintEvent* event) override {
-      QWidget::paintEvent(event);
-      auto painter = QPainter(this);
-      painter.setPen(QColor("#C8C8C8"));
-      painter.drawRect(0, 0, width() - 1, height() - 1);
-      painter.setPen("#4B23A0");
-      painter.drawLine(0, 0, static_cast<QWidget*>(parent())->width() - 1, 0);
-    }
-
-  private:
-    QWidget* m_widget;
-    DropShadow* m_shadow;
-
-    void move_to_parent() {
-      auto parent_widget = static_cast<QWidget*>(parent());
-      auto pos = parent_widget->window()->mapToGlobal(
-        parent_widget->frameGeometry().bottomLeft());
-      move(pos);
-      raise();
-    }
-};
-
-class DropDownItem : public QWidget {
-  public:
-
-    using SelectedSignal = Signal<void (const QVariant& value)>;
-
-    boost::signals2::connection connect_selected_signal(
-        const SelectedSignal::slot_type& slot) const {
-      return m_selected_signal.connect(slot);
-    }
-
-    DropDownItem(const QVariant& value,
-        QWidget* parent = nullptr)
-        : QWidget(parent),
-          m_value(value),
-          m_is_highlighted(false) {
-      setAttribute(Qt::WA_Hover);
-      auto font = QFont("Roboto");
-      font.setPixelSize(scale_height(12));
-      setFont(font);
-    }
-
-    const QVariant& get_value() const {
-      return m_value;
-    }
-
-    void set_highlight() {
-      m_is_highlighted = true;
-      update();
-    }
-
-    void reset_highlight() {
-      m_is_highlighted = false;
-      update();
-    }
-
-  protected:
-    void mousePressEvent(QMouseEvent* event) override {
-      if(event->button() == Qt::LeftButton) {
-        m_selected_signal(m_value);
-      }
-    }
-    void paintEvent(QPaintEvent* event) override {
-      auto painter = QPainter(this);
-      if(underMouse() || m_is_highlighted) {
-        painter.fillRect(rect(), QColor("#F2F2FF"));
-      } else {
-        painter.fillRect(rect(), Qt::white);
-      }
-      // TODO: move to anon namespace
-      auto LEFT_PADDING = []{ return scale_width(8); };
-      auto RIGHT_PADDING = []{ return scale_width(12); };
-      auto metrics = QFontMetrics(font());
-      auto shortened_text = metrics.elidedText(
-        m_item_delegate.displayText(m_value), Qt::ElideRight,
-        width() - RIGHT_PADDING());
-      painter.drawText(LEFT_PADDING(), metrics.height(), shortened_text);
-    }
-
-  private:
-    QVariant m_value;
-    mutable SelectedSignal m_selected_signal;
-    bool m_is_highlighted;
-    CustomVariantItemDelegate m_item_delegate;
-};
-
-class DropDownList : public DropDownWindow {
-  public:
-
-    using SelectedSignal = Signal<void (const QVariant& value)>;
-
-    boost::signals2::connection connect_selected_signal(
-        const SelectedSignal::slot_type& slot) const {
-      return m_selected_signal.connect(slot);
-    }
-
-    DropDownList(std::vector<DropDownItem*> items,
-        QWidget* parent = nullptr)
-        : DropDownWindow(parent),
-          m_max_displayed_items(5) {
-      //setFocusProxy(parent);
-      m_scroll_area = new ScrollArea(this);
-      m_scroll_area->setWidgetResizable(true);
-      auto main_widget = new QWidget(this);
-      m_layout = new QVBoxLayout(main_widget);
-      m_layout->setContentsMargins({});
-      m_layout->setSpacing(0);
-      m_scroll_area->setWidget(main_widget);
-      set_widget(m_scroll_area);
-      set_items(items);
-      parent->installEventFilter(this);
-    }
-
-    void set_items(std::vector<DropDownItem*> items) {
-      while(auto item = m_layout->takeAt(0)) {
-        delete item->widget();
-        delete item;
-      }
-      for(auto item : items) {
-        m_layout->addWidget(item);
-        item->connect_selected_signal([=] (const auto& value) {
-          on_item_selected(value);
-        });
-      }
-      if(m_layout->count() > 0) {
-        setFixedHeight(std::min(m_max_displayed_items, m_layout->count()) *
-          m_layout->itemAt(0)->widget()->height());
-      } else {
-        hide();
-      }
-    }
-
-  protected:
-    bool eventFilter(QObject* watched, QEvent* event) override {
-      if(watched == parent()) {
-        if(event->type() == QEvent::KeyPress) {
-          auto e = static_cast<QKeyEvent*>(event);
-          if(e->key() == Qt::Key_Up) {
-            focus_previous();
-            return true;
-          } else if(e->key() == Qt::Key_Down) {
-            focus_next();
-            return true;
-          } else if(e->key() == Qt::Key_Enter || e->key() == Qt::Key_Return) {
-            if(isVisible() && m_highlight_index) {
-              m_selected_signal(get_widget(*m_highlight_index)->get_value());
-            }
-          }
-        }
-      }
-      return DropDownWindow::eventFilter(watched, event);
-    }
-    void focusInEvent(QFocusEvent* event) override {
-      auto a = 0;
-    }
-    void hideEvent(QHideEvent* event) override {
-      m_scroll_area->verticalScrollBar()->setValue(0);
-      if(m_highlight_index) {
-        get_widget(*m_highlight_index)->reset_highlight();
-        m_highlight_index = boost::none;
-      }
-    }
-    void keyPressEvent(QKeyEvent* event) override {
-      if(event->key() == Qt::Key_Up) {
-        focus_previous();
-      } else if(event->key() == Qt::Key_Down) {
-        focus_next();
-      } else if(event->key() == Qt::Key_Escape) {
-        hide();
-      } else if(event->key() == Qt::Key_Enter) {
-        if(m_highlight_index) {
-          m_selected_signal(get_widget(*m_highlight_index)->get_value());
-          hide();
-        }
-      }
-      DropDownWindow::keyPressEvent(event);
-    }
-
-  private:
-    mutable SelectedSignal m_selected_signal;
-    int m_max_displayed_items;
-    QVBoxLayout* m_layout;
-    ScrollArea* m_scroll_area;
-    boost::optional<int> m_highlight_index;
-
-    DropDownItem* get_widget(int index) {
-      return static_cast<DropDownItem*>(m_layout->itemAt(index)->widget());
-    }
-    void focus_next() {
-      if(m_layout->count() == 0) {
-        return;
-      }
-      if(!m_highlight_index) {
-        set_highlight(0);
-        return;
-      }
-      set_highlight((*m_highlight_index) + 1);
-    }
-    void focus_previous() {
-      if(m_layout->count() == 0) {
-        return;
-      }
-      if(!m_highlight_index) {
-        set_highlight(m_layout->count() - 1);
-        return;
-      }
-      set_highlight((*m_highlight_index) - 1);
-    }
-    void set_highlight(int index) {
-      if(m_highlight_index) {
-        get_widget(*m_highlight_index)->reset_highlight();
-      }
-      if(index < 0) {
-        index = m_layout->count() - 1;
-      } else if(index > m_layout->count() - 1) {
-        index = 0;
-      }
-      m_highlight_index = index;
-      get_widget(*m_highlight_index)->set_highlight();
-      scroll_to_highlight();
-    }
-    void scroll_to_highlight() {
-      if(m_highlight_index != m_layout->count() - 1) {
-        m_scroll_area->ensureWidgetVisible(get_widget(*m_highlight_index), 0,
-          0);
-      } else {
-        m_scroll_area->verticalScrollBar()->setValue(
-          m_scroll_area->verticalScrollBar()->maximum());
-      }
-    }
-    void on_item_selected(const QVariant& value) {
-      m_selected_signal(value);
-      hide();
-    }
-};
-
-class DropDownMenu : public QWidget {
-  public:
-
-    using SelectedSignal = Signal<void (const QVariant& value)>;
-
-    boost::signals2::connection connect_selected_signal(
-        const SelectedSignal::slot_type& slot) const {
-      return m_selected_signal.connect(slot);
-    }
-
-    explicit DropDownMenu(const std::vector<QString>& items,
-        QWidget* parent)
-        : QWidget(parent),
-          m_dropdown_image(imageFromSvg(":/Icons/arrow-down.svg", scale(6, 4))) {
-      setFocusPolicy(Qt::StrongFocus);
-      if(!items.empty()) {
-        m_current_text = items.front();
-      }
-      set_items(items);
-      m_menu_list->connect_selected_signal([=] (const auto& value) {
-        on_item_selected(value);
-      });
-    }
-
-    void set_list_width(int width) {
-      m_menu_list->setFixedWidth(width);
-    }
-
-    void set_current_text(const QString& text) {
-      m_current_text = text;
-    }
-
-    void set_items(const std::vector<QString>& items) {
-      auto widget_items = std::vector<DropDownItem*>();
-      widget_items.reserve(items.size());
-      for(auto& item : items) {
-        auto item_widget = new DropDownItem(item, this);
-        item_widget->setFixedHeight(scale_height(20));
-        widget_items.push_back(item_widget);
-      }
-      m_menu_list = new DropDownList(widget_items, this);
-      m_menu_list->setFixedWidth(width());
-    }
-
-    const QString& get_text() const {
-      return m_current_text;
-    }
-
-  protected:
-    void mousePressEvent(QMouseEvent* event) override {
-      if(event->button() == Qt::LeftButton) {
-        if(!m_menu_list->isVisible()) {
-          m_menu_list->show();
-        } else {
-          m_menu_list->hide();
-        }
-      }
-    }
-    void paintEvent(QPaintEvent* event) override {
-      // TODO: move to anon namespace
-      auto PADDING = [] { return scale_width(8); };
-      auto painter = QPainter(this);
-      if(hasFocus() || isAncestorOf(focusWidget())) {
-        painter.fillRect(event->rect(), QColor("#4B23A0"));
-      } else {
-        painter.fillRect(event->rect(), QColor("#C8C8C8"));
-      }
-      painter.fillRect(1, 1, width() - 2, height() - 2, Qt::white);
-      auto font = QFont("Roboto");
-      font.setPixelSize(scale_height(12));
-      painter.setFont(font);
-      auto metrics = QFontMetrics(font);
-      painter.drawText(QPoint(PADDING(),
-        (height() / 2) + (metrics.ascent() / 2) - 1),
-        metrics.elidedText(m_current_text, Qt::ElideRight,
-        width() - (PADDING() * 3)));
-      painter.drawImage(
-        QPoint(width() - (m_dropdown_image.width() + PADDING()),
-        scale_height(11)), m_dropdown_image);
-    }
-    void resizeEvent(QResizeEvent* event) override {
-      m_menu_list->setFixedWidth(width());
-    }
-
-  private:
-    mutable SelectedSignal m_selected_signal;
-    QString m_current_text;
-    QImage m_dropdown_image;
-    DropDownList* m_menu_list;
-
-    void on_item_selected(const QVariant& value) {
-      m_current_text = value.value<QString>();
-      m_selected_signal(value);
-      update();
-    }
-};
+#include "Spire/Ui/ColorSelectorButton.hpp"
+#include "Spire/Ui/StaticDropDownMenu.hpp"
 
 int main(int argc, char** argv) {
   auto application = new QApplication(argc, argv);
@@ -438,13 +19,19 @@ int main(int argc, char** argv) {
   auto label = new QLabel("Null", test_window);
   label->setFont(QFont("Roboto", 14));
   label->move(scale_width(10), scale_height(10));
-  auto dropdown1 = new ::DropDownMenu({"One", "Two", "Three", "Four", "Five",
-    "Six", "Seven"}, test_window);
+  auto dropdown1 = new StaticDropDownMenu({"One", "Two", "Three", "Four",
+    "Five", "Six", "Seven"}, test_window);
   dropdown1->setFixedSize(scale(100, 28));
   dropdown1->connect_selected_signal([&] (const auto& value) {
     label->setText(value.value<QString>());
   });
   layout->addWidget(dropdown1);
+  auto c = new ColorSelectorButton(Qt::red, test_window);
+  c->setFixedSize(scale(100, 28));
+  c->connect_color_signal([&] (const auto& color) {
+    label->setText(color.name());
+  });
+  layout->addWidget(c);
   test_window->resize(scale(800, 150));
   test_window->show();
   dropdown1->setFocus();
