+++ Applications/Spire/Source/Ui/ColorSelectorButton.cpp
@@ -2,6 +2,7 @@
 #include <QEvent>
 #include <QKeyEvent>
 #include <QPainter>
+#include "Spire/Ui/DropDownWindow.hpp"
 
 using namespace boost::signals2;
 using namespace Spire;
@@ -11,12 +12,12 @@ ColorSelectorButton::ColorSelectorButton(const QColor& current_color,
     : QWidget(parent) {
   setFocusPolicy(Qt::StrongFocus);
   setAttribute(Qt::WA_Hover);
-  m_dropdown = new ColorSelectorDropDown(current_color, this);
-  m_dropdown->connect_color_signal([=] (const auto& color) {
+  m_selector_widget = new ColorSelectorDropDown(current_color, this);
+  m_selector_widget->connect_color_signal([=] (const auto& color) {
     on_color_selected(color);
   });
-  m_dropdown->hide();
-  window()->installEventFilter(this);
+  auto dropdown = new DropDownWindow(true, this);
+  dropdown->set_widget(m_selector_widget);
   set_color(current_color);
 }
 
@@ -26,7 +27,7 @@ const QColor& ColorSelectorButton::get_color() const {
 
 void ColorSelectorButton::set_color(const QColor& color) {
   m_current_color = color;
-  m_dropdown->set_color(color);
+  m_selector_widget->set_color(color);
   update();
 }
 
@@ -35,56 +36,36 @@ connection ColorSelectorButton::connect_color_signal(
   return m_color_signal.connect(slot);
 }
 
-bool ColorSelectorButton::eventFilter(QObject* watched, QEvent* event) {
-  if(watched == window()) {
-    if(event->type() == QEvent::Move) {
-      if(m_dropdown->isVisible()) {
-        move_color_dropdown();
-      }
-    } else if(event->type() == QEvent::WindowDeactivate &&
-        !m_dropdown->isActiveWindow()) {
-      hide_dropdown();
-    } else if(event->type() == QEvent::MouseButtonPress) {
-      hide_dropdown();
-    }
-  } else if(watched == m_dropdown) {
-    if(event->type() == QEvent::Hide) {
-      hide_dropdown();
-    }
-  }
-  return QWidget::eventFilter(watched, event);
-}
+//void ColorSelectorButton::focusOutEvent(QFocusEvent* event) {
+//  if(!m_dropdown->isActiveWindow()) {
+//    hide_dropdown();
+//  }
+//  update();
+//}
 
-void ColorSelectorButton::focusOutEvent(QFocusEvent* event) {
-  if(!m_dropdown->isActiveWindow()) {
-    hide_dropdown();
-  }
-  update();
-}
+//void ColorSelectorButton::keyPressEvent(QKeyEvent* event) {
+//  if(event->key() == Qt::Key_Escape) {
+//    if(m_dropdown->isVisible()) {
+//      m_dropdown->hide();
+//    }
+//  }
+//}
 
-void ColorSelectorButton::keyPressEvent(QKeyEvent* event) {
-  if(event->key() == Qt::Key_Escape) {
-    if(m_dropdown->isVisible()) {
-      m_dropdown->hide();
-    }
-  }
-}
-
-void ColorSelectorButton::mousePressEvent(QMouseEvent* event) {
-  if(m_dropdown->isVisible()) {
-    hide_dropdown();
-  } else {
-    move_color_dropdown();
-    m_dropdown->show();
-    m_dropdown->activateWindow();
-  }
-}
+//void ColorSelectorButton::mousePressEvent(QMouseEvent* event) {
+//  if(m_dropdown->isVisible()) {
+//    hide_dropdown();
+//  } else {
+//    move_color_dropdown();
+//    m_dropdown->show();
+//    m_dropdown->activateWindow();
+//  }
+//}
 
 void ColorSelectorButton::paintEvent(QPaintEvent* event) {
   QWidget::paintEvent(event);
   auto painter = QPainter(this);
   painter.fillRect(0, 0, 1000, 1000, Qt::red);
-  if(hasFocus() || m_dropdown->isActiveWindow() || underMouse()) {
+  if(hasFocus() || m_selector_widget->isActiveWindow() || underMouse()) {
     painter.fillRect(event->rect(), QColor("#4B23A0"));
   } else {
     painter.fillRect(event->rect(), QColor("#C8C8C8"));
@@ -94,18 +75,18 @@ void ColorSelectorButton::paintEvent(QPaintEvent* event) {
   painter.fillRect(2, 2, width() - 4, height() - 4, m_current_color);
 }
 
-void ColorSelectorButton::hide_dropdown() {
-  m_dropdown->hide();
-}
+//void ColorSelectorButton::hide_dropdown() {
+//  m_dropdown->hide();
+//}
 
-void ColorSelectorButton::move_color_dropdown() {
-  auto x_pos = static_cast<QWidget*>(parent())->mapToGlobal(
-    geometry().bottomLeft()).x();
-  auto y_pos = static_cast<QWidget*>(parent())->mapToGlobal(
-    frameGeometry().bottomLeft()).y();
-  m_dropdown->move(x_pos, y_pos + 1);
-  m_dropdown->raise();
-}
+//void ColorSelectorButton::move_color_dropdown() {
+//  auto x_pos = static_cast<QWidget*>(parent())->mapToGlobal(
+//    geometry().bottomLeft()).x();
+//  auto y_pos = static_cast<QWidget*>(parent())->mapToGlobal(
+//    frameGeometry().bottomLeft()).y();
+//  m_dropdown->move(x_pos, y_pos + 1);
+//  m_dropdown->raise();
+//}
 
 void ColorSelectorButton::on_color_selected(const QColor& color) {
   m_current_color = color;
