+++ Applications/Spire/Source/Ui/ColorSelectorDropDown.cpp
@@ -68,17 +68,15 @@ namespace {
 
 ColorSelectorDropDown::ColorSelectorDropDown(const QColor& current_color,
     QWidget* parent)
-    : QWidget(parent, Qt::Tool | Qt::FramelessWindowHint),
+    : QWidget(parent),
       m_recent_colors(RecentColors::get_instance()),
       m_current_color(current_color),
       m_original_color(current_color) {
-  m_drop_shadow = new DropShadow(true, false, this);
   setFixedSize(scale(232, 198));
   setObjectName("color_selector_drop_down");
   setStyleSheet(R"(
     #color_selector_drop_down {
       background-color: #FFFFFF;
-      border: 1px solid #C8C8C8;
     })");
   auto layout = new QVBoxLayout(this);
   layout->setContentsMargins(HORIZONTAL_PADDING(), VERTICAL_PADDING(),
@@ -186,17 +184,15 @@ void ColorSelectorDropDown::childEvent(QChildEvent* event) {
 bool ColorSelectorDropDown::eventFilter(QObject* watched, QEvent* event) {
   if(event->type() == QEvent::KeyPress) {
     auto e = static_cast<QKeyEvent*>(event);
-    if(e->key() == Qt::Key_Escape) {
-      hide();
-    } else if(e->key() == Qt::Key_Enter || e->key() == Qt::Key_Return) {
+    if(e->key() == Qt::Key_Enter || e->key() == Qt::Key_Return) {
       on_color_selected(m_current_color);
-      hide();
+      window()->hide();
     }
   } else if(event->type() == QEvent::MouseButtonDblClick) {
     auto e = static_cast<QMouseEvent*>(event);
     if(e->button() == Qt::LeftButton) {
       on_color_selected(m_current_color);
-      hide();
+      window()->hide();
     }
   }
   return QWidget::eventFilter(watched, event);
